validates the scheduler
