+++ api/v1alpha1/iochaos_webhook.go
@@ -87,7 +87,7 @@ func (in *IoChaos) Validate() error {
 
 // ValidateScheduler validates the scheduler and duration
 func (in *IoChaos) ValidateScheduler(spec *field.Path) field.ErrorList {
-	return ValidateScheduler(in.Spec.Duration, in.Spec.Scheduler, spec)
+	return ValidateScheduler(in, spec)
 }
 
 // ValidatePodMode validates the value with podmode
