+++ api/v1alpha1/kernelchaos_webhook.go
@@ -82,7 +82,7 @@ func (in *KernelChaos) Validate() error {
 
 // ValidateScheduler validates the scheduler and duration
 func (in *KernelChaos) ValidateScheduler(spec *field.Path) field.ErrorList {
-	return ValidateScheduler(in.Spec.Duration, in.Spec.Scheduler, spec)
+	return ValidateScheduler(in, spec)
 }
 
 // ValidatePodMode validates the value with podmode
