+++ api/v1alpha1/networkchaos_webhook.go
@@ -126,7 +126,7 @@ func (in *NetworkChaos) Validate() error {
 
 // ValidateScheduler validates the scheduler and duration
 func (in *NetworkChaos) ValidateScheduler(spec *field.Path) field.ErrorList {
-	return ValidateScheduler(in.Spec.Duration, in.Spec.Scheduler, spec)
+	return ValidateScheduler(in, spec)
 }
 
 // ValidatePodMode validates the value with podmode
